+++ Xen/mouse/mouse.h
@@ -280,8 +280,8 @@ public:
      */
     MouseThread(
         int  resolution,
-        int  fovX,
-        int  fovY,
+        double fovX,
+        double fovY,
         bool auto_shoot,
         float bScope_multiplier,
         IMouseInput* mouseInputDevice = nullptr
@@ -293,8 +293,8 @@ public:
      */
     void updateConfig(
         int resolution,
-        int fovX,
-        int fovY,
+        double fovX,
+        double fovY,
         bool auto_shoot,
         float bScope_multiplier
     );
